+++ src/themes/default/components/core/blocks/Microcart/Microcart.vue
@@ -105,6 +105,11 @@
           {{ $t('Go to checkout') }}
         </button-full>
         <instant-checkout v-if="isInstantCheckoutRegistered" class="no-outline button-full block brdr-none w-100 px10 py20 bg-cl-mine-shaft :bg-cl-th-secondary ripple weight-400 h4 cl-white sans-serif fs-medium mt20" />
+        <button-full
+          @click.native="clearCart"
+        >
+          Clear cart
+        </button-full>
       </div>
     </div>
   </div>
@@ -184,6 +189,19 @@ export default {
     },
     onEscapePress () {
       this.closeMicrocart()
+    },
+    clearCart () {
+      this.$store.dispatch('notification/spawnNotification', {
+        type: 'warning',
+        message: i18n.t('Are you sure you would like to remove all the items from the shopping cart?'),
+        action1: { label: i18n.t('OK'),
+          action: () => {
+            this.$store.dispatch('cart/clear')
+          }
+        },
+        action2: { label: i18n.t('Cancel'), action: 'close' },
+        hasNoTimeout: true
+      })
     }
   }
 }